more pricicjon

+++ my_lib/include/StackedBarModel.hpp
@@ -7,14 +7,14 @@ struct Data
 {
     std::string name;
     Color col;
-    float val;
+    double val;
 };
 
 class StackedBarModel
 {
 private:
     std::vector<Data> data;
-    float total_volume = 0;
+    double total_volume = 0;
 
 public:
     StackedBarModel() = default;
@@ -23,9 +23,9 @@ public:
     StackedBarModel &operator=(const StackedBarModel &other) = default;
 
     void reset(void);
-    void add_value(const std::string &name, Color col, float val);
+    void add_value(const std::string &name, Color col, double val);
     const std::vector<Data> &get_data(void) const { return data; }
-    float get_total_volume(void) const { return total_volume; }
+    double get_total_volume(void) const { return total_volume; }
 
     Data *operator[](const std::string& name);
     bool contains(const std::string& name) const;
